requirements.txt
streamlit==1.37.1
httpx==0.27.0
pydantic==2.8.2
jsonschema==4.23.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
Jinja2==3.1.4
Pillow==10.4.0
python-dateutil==2.9.0.post0
cryptography==43.0.0
PyYAML==6.0.2
openpyxl==3.1.5
python-docx==1.1.2
pypdfium2==4.30.0
numpy<2.0
pyarrow>=15.0.0
requests>=2.31.0
watchdog